2008-08-21  Sancho Lerena <slerena@gmail.com>
	* pandoradb*.sql: tplanned_downtime: Fixed a typo, changed timestamp 
	types, added executed field and removed id_agent_module.

	* manage_config.php:  (Problem copying keep_alive modules
	and alerts and solved bad message with no data supplied.

	* manage_config_remote.php: Solved BUG #2042788 Fixed, was missing 
	extension in filename check.

	* planned_downtime.php: Major rewrite. Now works with server model of
	planned downtimes based on select agents to disable in specific times.
	Using unixtime was a very nice idea, so the javascript controls are 
	fantastic to do this job!. Thanks Esteban and Evi!.

	* manage_recontask*: Some issues with styles, using new div for button.

	* modificar_server.php:  Some issues with styles, using new div for button

	* config.inc.php: By default, show errors, easy to debug first problems.

	* functions_db.php: Fixed problems detected in borrar_incidencia() function
	(this is delete_incident in spanish).

	* estado_agente.php: Added support to search by IP address in agent view.

	* networkmap.php: Regen on by default, Font label shortened.



git-svn-id: https://svn.code.sf.net/p/pandora/code/trunk@1019 c3f86ba8-e40f-0410-aaad-9ba5e7f4b01f

+++ pandora_console/include/functions_db.php
@@ -767,26 +767,30 @@ function dame_generic_string_data ($id) {
 function borrar_incidencia ($id_inc) {
 	global $config;
 	
-	$sql = sprintf ("DELETE FROM `tincidencia` WHERE `id_incidencia` = %d",$id_inc);
+	$sql = sprintf ("DELETE FROM `tincidencia` WHERE `id_incidencia` = %d", $id_inc);
 	process_sql ($sql);
-	$sql = sprintf ("SELECT `id_nota` FROM `tnota_inc` WHERE `id_incidencia` = %d".$id_inc);
+	$sql = sprintf ("SELECT `id_nota` FROM `tnota_inc` WHERE `id_incidencia` = %d ", $id_inc);
 	$rows = get_db_all_rows_sql ($sql);
-	foreach ($rows as $row) {
-		$sql = sprintf ("DELETE FROM `tnota` WHERE `id_nota` = %d",$row["id_nota"]);
+	if ($rows){
+		foreach ($rows as $row) {
+			$sql = sprintf ("DELETE FROM `tnota` WHERE `id_nota` = %d",$row["id_nota"]);
+			process_sql ($sql);
+		}
+		$sql = "DELETE FROM `tnota_inc` WHERE `id_incidencia` = $id_inc";
 		process_sql ($sql);
 	}
-	$sql = "DELETE FROM `tnota_inc` WHERE `id_incidencia` = ".$id_inc;
-	process_sql ($sql);
 	
 	// Delete attachments
-	$sql = sprintf ("SELECT `id_attachment`,`filename` FROM `tattachment` WHERE `id_incidencia` = %d",$id_inc);
+	$sql = sprintf ("SELECT `id_attachment`,`filename` FROM `tattachment` WHERE `id_incidencia` = %d", $id_inc);
 	$rows = get_db_all_rows_sql ($sql);
-	foreach ($rows as $row) {
-		// Unlink all attached files for this incident
-		unlink ($attachment_store."attachment/pand".$row["id_attachment"]."_".$row["filename"]);
+	if ($rows){
+		foreach ($rows as $row) {
+			// Unlink all attached files for this incident
+			unlink ($attachment_store."attachment/pand".$row["id_attachment"]."_".$row["filename"]);
+		}
+		$sql = sprintf ("DELETE FROM `tattachment` WHERE `id_incidencia` = %d",$id_inc);
+		process_sql ($sql);
 	}
-	$sql = sprintf ("DELETE FROM `tattachment` WHERE `id_incidencia` = %d",$id_inc);
-	process_sql ($sql);
 }
 
 /**
